🚧 Add ebo command to ServerGUI and refactor egg handling in GameDataManager
