+++ zappy_gui_src/DataManager/GameTile.cpp
@@ -12,8 +12,7 @@ GameTile::GameTile(int xCoord, int yCoord)
 : x(xCoord), y(yCoord) {}
 
 GameTile::GameTile(GameTile &&other) noexcept
-: x(other.x), y(other.y), tileMesh(std::move(other.tileMesh)),
-eggs(std::move(other.eggs)) {}
+: x(other.x), y(other.y), tileMesh(std::move(other.tileMesh)) {}
 
 GameTile &GameTile::operator=(GameTile &&other) noexcept {
     if (this != &other) {
@@ -21,7 +20,6 @@ GameTile &GameTile::operator=(GameTile &&other) noexcept {
         x = other.x;
         y = other.y;
         tileMesh = std::move(other.tileMesh);
-        eggs = std::move(other.eggs);
     }
     return *this;
 }
@@ -59,18 +57,6 @@ Vec3d GameTile::getWorldPos() const {
     if (!tileMesh)
         return Vec3d(0, 0, 0);
     return tileMesh->getPosition();
-}
-
-const std::vector<GameTile::Egg> &GameTile::getEggs() const {
-    return eggs;
-}
-
-void GameTile::addEgg(int id, int team) {
-    std::lock_guard<std::mutex> lock(mutexDatas);
-    Vec3d position = getWorldPos();
-    position.Y += 0.2f;
-    eggs.emplace_back(id, team, MeshImporter::i().importMesh("DroneEgg",
-        position, Vec3d(0.2f)));
 }
 
 void GameTile::setRessources(int food, int r1, int r2, int r3, int r4,
@@ -120,12 +106,4 @@ float offsetX, float offsetY, float offsetZ) {
     }
 }
 
-// ======= Egg ========= //
-
-GameTile::Egg::Egg(int id, int team,
-const std::shared_ptr<irr::scene::IAnimatedMeshSceneNode> &eggMesh)
-: id(id), team(team), EggMesh(eggMesh) {
-    if (!EggMesh)
-        throw GUI::ShaderCompilationException("Error creating egg mesh");
-}
 }  // namespace GUI
